Fixing SQL Scan Issues
Improving SQL code generation by checking when table has one-to-many relationship vs. when it does
not, and with this information either append or prepend the table reference, therefore the query does
not fail on not being able to find tables/columns.

Adding a SQL formatter to log queries, making sure we can quickly troubleshoot them. Furthermore
adding more tests, and removing a couple of FIXMEs.